+++ node-api/Models/LinkDisconnectionEvent.cs
@@ -14,7 +14,7 @@ public record LinkDisconnectionEvent : NetworkEventDatagram
     /// Timestamp (secs since 1/1/70) (Optional)
     /// </summary>
     [JsonPropertyName("time")]
-    public long? TimeUnixSeconds { get; init; }
+    public decimal? TimeUnixSeconds { get; init; }
 
     /// <summary>
     /// Callsign of reporting node (Required)
