+++ node-api/Models/NodeStatusReportEvent.cs
@@ -16,7 +16,7 @@ public record NodeStatusReportEvent : NetworkEventDatagram
     /// Timestamp (secs since 1/1/70) (Optional)
     /// </summary>
     [JsonPropertyName("time")]
-    public long? TimeUnixSeconds { get; init; }
+    public decimal? TimeUnixSeconds { get; init; }
 
     /// <summary>
     /// Node Callsign (Required)
